M1-40
- add initial implementation for image node logging to test format on robot
